Refactor: move OpenGl wrapper code from Drawing to drawLib
