+++ SourceCode/AgOpenGPS.Core/DrawLib/Texture2D.cs
@@ -3,7 +3,7 @@ using System.Drawing.Imaging;
 using AgOpenGPS.Core.Models;
 using OpenTK.Graphics.OpenGL;
 
-namespace AgOpenGPS.Core.Drawing
+namespace AgOpenGPS.Core.DrawLib
 {
     public class Texture2D
     {
